# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zero-Click CRM API
  version: 1.0.0
  description: >-
    Documentación de la API para el backend Zero-Click CRM. Incluye ingestión multicanal,
    procesamiento de audio/video, consultas CRM, chat asistido por IA, alertas, notas,
    knowledge base y endpoints administrativos.
  contact:
    name: Equipo Relay
servers:
  - url: http://localhost:4000
    description: Servidor de desarrollo local
  - url: https://your-deployment-url.example
    description: Servidor de producción (actualiza esta URL al desplegar)
...
paths:
...
  /api/crm/contacts:
    get:
      summary: Lista contactos del CRM
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: search
          in: query
          schema:
            type: string
        - name: sentiment
          in: query
          schema:
            $ref: '#/components/schemas/Sentiment'
        - name: personKind
          in: query
          schema:
            type: string
            enum: [employee, client, supplier, partner, other]
        - name: isClient
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/UpdatedAfterQuery'
        - $ref: '#/components/parameters/UpdatedBeforeQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Lista paginada de contactos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactsListResponse'
  /api/crm/companies:
    get:
      summary: Lista compañías
      tags: [CRM]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Lista paginada de compañías
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniesListResponse'
  /api/crm/interactions:
    get:
      summary: Lista interacciones
      tags: [CRM]
      parameters:
        - name: contactId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: channel
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: minBudget
          in: query
          schema:
            type: number
        - name: maxBudget
          in: query
          schema:
            type: number
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Lista paginada de interacciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionsListResponse'
  /api/crm/work-items:
    get:
      summary: Lista work items
      tags: [CRM]
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WorkItemStatus'
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: assigneeId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: dueBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: dueAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: onlyOverdue
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Lista paginada de work items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemsListResponse'
  /api/crm/fresh-data:
    get:
      summary: Lista señales fresh data
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: topic
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Lista paginada de señales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreshDataListResponse'
  /api/crm/insights/summary:
    get:
      summary: Dashboard resumen por compañía
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Resumen con métricas, deals y próximos pasos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightsSummaryResponse'
  /api/crm/timeline:
    get:
      summary: Timeline combinado de interacciones, work items y señales
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: contactId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista unificada ordenada por fecha
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
  /api/crm/companies/{id}/overview:
    get:
      summary: Overview completo de una compañía
      tags: [CRM]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: interactionsLimit
          in: query
          schema:
            type: integer
        - name: workItemsLimit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Información consolidada de la compañía
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyOverviewResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/crm/contacts/{id}/overview:
    get:
      summary: Overview completo de un contacto
      tags: [CRM]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: interactionsLimit
          in: query
          schema:
            type: integer
        - name: workItemsLimit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Información consolidada del contacto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactOverviewResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/crm/insights/trends:
    get:
      summary: Tendencias de interacciones, work items y fresh data
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Datos agrupados por día
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResponse'
  /api/crm/insights/actionable:
    get:
      summary: Insights accionables (contactos en riesgo, alertas, tareas vencidas)
      tags: [CRM]
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200':
          description: Información lista para actuar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionableInsightsResponse'
...
  /api/work-items:
    post:
      summary: Crear work item manualmente
      tags: [WorkItems]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemCreateRequest'
      responses:
        '200':
          description: Work item creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
...
    Sentiment:
      type: string
      enum: [positive, neutral, negative]
    Priority:
      type: string
      enum: [low, medium, high, critical]
    WorkItemStatus:
      type: string
      enum: [pending, in_progress, blocked, completed]
    AlertSeverity:
      type: string
      enum: [low, medium, high, critical]
...
    ContactsListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        sentiment:
          $ref: '#/components/schemas/Sentiment'
        person_kind:
          type: string
        is_client:
          type: boolean
        updated_at:
          type: string
          format: date-time
        company:
          $ref: '#/components/schemas/Company'
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        domain:
          type: string
          nullable: true
    CompaniesListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Company'
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    InteractionsListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Interaction:
      type: object
      properties:
        id:
          type: string
        channel:
          type: string
        occurred_at:
          type: string
          format: date-time
        notes:
          type: string
        participants:
          type: array
          items:
            type: string
        budget:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        requirements:
          type: object
          nullable: true
        kpis:
          type: object
          nullable: true
        data:
          type: object
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        contact:
          $ref: '#/components/schemas/ContactBasic'
        company:
          $ref: '#/components/schemas/Company'
    ContactBasic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          nullable: true
        sentiment:
          $ref: '#/components/schemas/Sentiment'
          nullable: true
    WorkItemsListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    WorkItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/WorkItemStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        due_date:
          type: string
          format: date-time
          nullable: true
        owner:
          $ref: '#/components/schemas/ContactBasic'
        assignee:
          $ref: '#/components/schemas/ContactBasic'
        company:
          $ref: '#/components/schemas/Company'
    FreshDataListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/FreshData'
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    FreshData:
      type: object
      properties:
        id:
          type: string
        topic:
          type: string
        source:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        detected_at:
          type: string
          format: date-time
          nullable: true
        company:
          $ref: '#/components/schemas/Company'
    InsightsSummaryResponse:
      type: object
      properties:
        ok:
          type: boolean
        summary:
          type: object
        workItemsStatus:
          type: array
          items:
            type: object
        sentiment:
          type: array
          items:
            type: object
        topDeals:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        upcomingWorkItems:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
        recentInteractions:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
    TimelineEntry:
      type: object
      properties:
        type:
          type: string
          enum: [interaction, work_item, fresh_data]
        id:
          type: string
        occurredAt:
          type: string
          format: date-time
        sortDate:
          type: string
          format: date-time
        channel:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/WorkItemStatus'
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        topic:
          type: string
          nullable: true
        company:
          $ref: '#/components/schemas/Company'
          nullable: true
        contact:
          $ref: '#/components/schemas/ContactBasic'
          nullable: true
    TimelineResponse:
      type: object
      properties:
        ok:
          type: boolean
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEntry'
        total:
          type: integer
    CompanyOverviewResponse:
      type: object
      properties:
        ok:
          type: boolean
        company:
          $ref: '#/components/schemas/Company'
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        workItems:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
        interactions:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        freshData:
          type: array
          items:
            $ref: '#/components/schemas/FreshData'
        pipeline:
          type: object
        sentiment:
          type: array
          items:
            type: object
        workItemsStatus:
          type: array
          items:
            type: object
    ContactOverviewResponse:
      type: object
      properties:
        ok:
          type: boolean
        contact:
          $ref: '#/components/schemas/Contact'
        interactions:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        workItems:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
        freshData:
          type: array
          items:
            $ref: '#/components/schemas/FreshData'
        pipeline:
          type: object
        workItemsStatus:
          type: array
          items:
            type: object
    TrendsResponse:
      type: object
      properties:
        ok:
          type: boolean
        interactions:
          type: array
          items:
            type: object
        workItems:
          type: array
          items:
            type: object
        freshData:
          type: array
          items:
            type: object
    ActionableInsightsResponse:
      type: object
      properties:
        ok:
          type: boolean
        risky_contacts:
          type: array
          items:
            type: object
        open_alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        overdue_work_items:
          type: array
          items:
            $ref: '#/components/schemas/WorkItem'
        stale_companies:
          type: array
          items:
            type: object
    Alert:
      type: object
      properties:
        id:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        status:
          type: string
        message:
          type: string
        data:
          type: object
        company:
          $ref: '#/components/schemas/Company'
        contact:
          $ref: '#/components/schemas/ContactBasic'
...
    WorkItemCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        companyId:
          type: string
          nullable: true
        assigneeContactId:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
          default: medium
        data:
          type: object
          default: {}
    WorkItemCreateResponse:
      type: object
      properties:
        ok:
          type: boolean
        workItem:
          $ref: '#/components/schemas/WorkItem'
...
  parameters:
    CompanyIdQuery:
      name: companyId
      in: query
      schema:
        type: string
      description: Filtra por UUID de la compañía
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
    UpdatedAfterQuery:
      name: updatedAfter
      in: query
      schema:
        type: string
        format: date-time
    UpdatedBeforeQuery:
      name: updatedBefore
      in: query
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
                example: false
              error:
                type: string
...